revert customer account back to 2023-07
