+++ packages/ui-extensions-react/src/surfaces/customer-account.ts
@@ -6,23 +6,77 @@ export {
   Style,
 } from '@shopify/ui-extensions/customer-account';
 export type {
-  ApiForExtension,
+  AccessibilityRole,
+  Alignment,
+  AllowedComponentsForRenderExtension,
+  ApiForRenderExtension,
+  Appearance,
+  AtLeastOne,
   ArgumentsForExtension,
+  Autocomplete,
+  AutocompleteField,
+  AutocompleteGroup,
+  Background,
+  BackgroundFit,
+  BackgroundPosition,
+  BackgroundProps,
+  BackgroundRepeat,
+  BlockAlignment,
+  BorderRadius,
+  BorderProps,
+  CornerProps,
+  CornerRadius,
+  BorderStyle,
+  BorderWidth,
+  Breakpoint,
+  Columns,
+  Conditions,
+  ConditionalStyle,
+  ConditionalValue,
+  DayString,
+  DateString,
+  DateRange,
+  DisclosureActivatorProps,
+  DisclosureOpen,
+  DisabledDate,
   ExtensionTarget,
   ExtensionTargets,
-  RenderExtension,
-  RenderExtensionTarget,
-  RunExtension,
-  RunExtensionTarget,
-  ReturnTypeForExtension,
-  StandardApi,
-  Language,
-  Localization,
+  Fit,
+  GridItemSize,
   I18nTranslate,
-  I18n,
-  FullPageApi,
+  InlineAlignment,
+  MaybeConditionalStyle,
+  MaybeResponsiveConditionalStyle,
+  InteractiveConditions,
+  MaybeShorthandProperty,
+  NonPresentationalAccessibilityRole,
+  OverlayActivatorProps,
+  PopoverPosition,
+  RenderExtension,
+  RenderExtensions,
+  ResolutionCondition,
+  ReturnTypeForExtension,
+  Rows,
+  ScrollViewEvent,
+  ShopifyGlobal,
+  ShorthandProperty,
+  SelectedDate,
+  Size,
+  SizingProps,
+  Spacing,
+  SpacingProps,
+  Status,
+  Version,
+  ViewCoordinate,
+  ViewTranslate,
+  ViewportSizeCondition,
+  ViewPosition,
+  ViewPositionType,
+  ViewLikeAccessibilityRole,
+  YearMonth,
 } from '@shopify/ui-extensions/customer-account';
 
-export * from './customer-account/components';
 export * from './customer-account/hooks';
 export {render, reactExtension} from './customer-account/render';
+
+export * from './checkout/components';
