--- packages/ui-extensions-react/src/surfaces/customer-account/components/Card/index.ts
@@ -1,2 +1,0 @@
-export {Card} from './Card';
-export type {CardProps} from './Card';
