--- packages/ui-extensions-react/src/surfaces/customer-account/components/CustomerAccountAction/CustomerAccountAction.ts
@@ -1,14 +1,0 @@
-import {CustomerAccountAction as BaseCustomerAccountAction} from '@shopify/ui-extensions/customer-account';
-import {
-  createRemoteReactComponent,
-  ReactPropsFromRemoteComponentType,
-} from '@remote-ui/react';
-
-export type CustomerAccountActionProps = ReactPropsFromRemoteComponentType<
-  typeof BaseCustomerAccountAction
->;
-
-export const CustomerAccountAction = createRemoteReactComponent(
-  BaseCustomerAccountAction,
-  {fragmentProps: ['primaryAction', 'secondaryAction']},
-);
