--- packages/ui-extensions-react/src/surfaces/customer-account/components/CustomerAccountAction/index.ts
@@ -1,2 +1,0 @@
-export {CustomerAccountAction} from './CustomerAccountAction';
-export type {CustomerAccountActionProps} from './CustomerAccountAction';
