--- packages/ui-extensions-react/src/surfaces/customer-account/components/PaymentIcon/PaymentIcon.ts
@@ -1,11 +1,0 @@
-import {PaymentIcon as BasePaymentIcon} from '@shopify/ui-extensions/customer-account';
-import {
-  createRemoteReactComponent,
-  ReactPropsFromRemoteComponentType,
-} from '@remote-ui/react';
-
-export type PaymentIconProps = ReactPropsFromRemoteComponentType<
-  typeof BasePaymentIcon
->;
-
-export const PaymentIcon = createRemoteReactComponent(BasePaymentIcon);
