--- packages/ui-extensions-react/src/surfaces/customer-account/components/PaymentIcon/index.ts
@@ -1,2 +1,0 @@
-export {PaymentIcon} from './PaymentIcon';
-export type {PaymentIconProps} from './PaymentIcon';
