--- packages/ui-extensions-react/src/surfaces/customer-account/components/Thumbnail/index.ts
@@ -1,2 +1,0 @@
-export {Thumbnail} from './Thumbnail';
-export type {ThumbnailProps} from './Thumbnail';
