+++ packages/ui-extensions-react/src/surfaces/customer-account/hooks/buyer-identity.ts
@@ -1,9 +1,7 @@
 import type {
   Customer,
-  OrderStatusPurchasingCompany,
-  PurchasingCompany as CustomerAccountPurchasingCompany,
+  PurchasingCompany,
   RenderExtensionTarget,
-  RenderOrderStatusExtensionTarget,
 } from '@shopify/ui-extensions/customer-account';
 
 import {ScopeNotGrantedError} from '../errors';
@@ -17,7 +15,7 @@ import {useSubscription} from './subscription';
  * The value is `undefined` if the buyer isn't a known customer for this shop or if they haven't logged in yet.
  */
 export function useCustomer<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): Customer | undefined {
   const buyerIdentity = useApi<Target>().buyerIdentity;
 
@@ -35,7 +33,7 @@ export function useCustomer<
  * The value is `undefined` if the app does not have access to customer data.
  */
 export function useEmail<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): string | undefined {
   const buyerIdentity = useApi<Target>().buyerIdentity;
 
@@ -53,7 +51,7 @@ export function useEmail<
  * The value is `undefined` if the app does not have access to customer data.
  */
 export function usePhone<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): string | undefined {
   const buyerIdentity = useApi<Target>().buyerIdentity;
 
@@ -66,10 +64,6 @@ export function usePhone<
   return useSubscription(buyerIdentity.phone);
 }
 
-type PurchasingCompany<Target> = Target extends RenderOrderStatusExtensionTarget
-  ? OrderStatusPurchasingCompany | undefined
-  : CustomerAccountPurchasingCompany | undefined;
-
 /**
  * Provides information about the company and its location that the business customer
  * is purchasing on behalf of during a B2B checkout. It includes details that can be utilized to
@@ -79,7 +73,7 @@ type PurchasingCompany<Target> = Target extends RenderOrderStatusExtensionTarget
  */
 export function usePurchasingCompany<
   Target extends RenderExtensionTarget = RenderExtensionTarget,
->(): PurchasingCompany<Target> {
+>(): PurchasingCompany | undefined {
   const buyerIdentity = useApi<Target>().buyerIdentity;
 
   if (!buyerIdentity) {
@@ -88,7 +82,5 @@ export function usePurchasingCompany<
     );
   }
 
-  return useSubscription(
-    buyerIdentity.purchasingCompany,
-  ) as PurchasingCompany<Target>;
+  return useSubscription(buyerIdentity.purchasingCompany);
 }
