+++ packages/ui-extensions-react/src/surfaces/customer-account/hooks/cart-lines.ts
@@ -1,6 +1,6 @@
 import type {
   CartLine,
-  RenderOrderStatusExtensionTarget,
+  RenderExtensionTarget,
 } from '@shopify/ui-extensions/customer-account';
 
 import {useApi} from './api';
@@ -11,7 +11,7 @@ import {useSubscription} from './subscription';
  * your component if line items are added, removed, or updated.
  */
 export function useCartLines<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): CartLine[] {
   const {lines} = useApi<Target>();
