+++ packages/ui-extensions-react/src/surfaces/customer-account/hooks/checkout-settings.ts
@@ -1,6 +1,6 @@
 import type {
   CheckoutSettings,
-  RenderOrderStatusExtensionTarget,
+  RenderExtensionTarget,
 } from '@shopify/ui-extensions/customer-account';
 
 import {useApi} from './api';
@@ -10,7 +10,7 @@ import {useSubscription} from './subscription';
  * Returns the `checkoutSettings` applied to the checkout.
  */
 export function useCheckoutSettings<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): CheckoutSettings {
   return useSubscription(useApi<Target>().checkoutSettings);
 }
