+++ packages/ui-extensions-react/src/surfaces/customer-account/hooks/country.ts
@@ -1,6 +1,6 @@
 import {
   Country,
-  RenderOrderStatusExtensionTarget,
+  RenderExtensionTarget,
 } from '@shopify/ui-extensions/customer-account';
 
 import {useApi} from './api';
@@ -11,7 +11,7 @@ import {useSubscription} from './subscription';
  * your component if the country changes.
  */
 export function useLocalizationCountry<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): Country | undefined {
   const {localization} = useApi<Target>();
