--- packages/ui-extensions-react/src/surfaces/customer-account/hooks/i18n.ts
@@ -1,15 +1,0 @@
-import {
-  I18n,
-  RenderCustomerAccountExtensionTarget,
-} from '@shopify/ui-extensions/customer-account';
-import {useApi} from './api';
-
-/**
- Returns utilities for translating content and formatting values
-    according to the current localization of the user.
- */
-export function useI18n<
-  Target extends RenderCustomerAccountExtensionTarget = RenderCustomerAccountExtensionTarget,
->(): I18n {
-  return useApi<Target>().i18n;
-}
