+++ packages/ui-extensions-react/src/surfaces/customer-account/hooks/note.ts
@@ -1,4 +1,4 @@
-import type {RenderOrderStatusExtensionTarget} from '@shopify/ui-extensions/customer-account';
+import type {RenderExtensionTarget} from '@shopify/ui-extensions/customer-account';
 
 import {useApi} from './api';
 import {useSubscription} from './subscription';
@@ -7,7 +7,7 @@ import {useSubscription} from './subscription';
  * Returns the proposed `note` applied to the checkout.
  */
 export function useNote<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): string | undefined {
   return useSubscription(useApi<Target>().note);
 }
