+++ packages/ui-extensions-react/src/surfaces/customer-account/hooks/session-token.ts
@@ -1,6 +1,6 @@
 import {
   SessionToken,
-  RenderOrderStatusExtensionTarget,
+  RenderExtensionTarget,
 } from '@shopify/ui-extensions/customer-account';
 
 import {useApi} from './api';
@@ -9,7 +9,7 @@ import {useApi} from './api';
  * Provides access to session tokens, which can be used to verify token claims on your app's server.
  */
 export function useSessionToken<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): SessionToken {
   return useApi<Target>().sessionToken;
 }
