+++ packages/ui-extensions-react/src/surfaces/customer-account/hooks/shop.ts
@@ -1,5 +1,5 @@
 import type {
-  RenderOrderStatusExtensionTarget,
+  RenderExtensionTarget,
   Shop,
 } from '@shopify/ui-extensions/customer-account';
 
@@ -9,7 +9,7 @@ import {useApi} from './api';
  * Returns the `Shop` where the checkout is taking place.
  */
 export function useShop<
-  Target extends RenderOrderStatusExtensionTarget = RenderOrderStatusExtensionTarget,
+  Target extends RenderExtensionTarget = RenderExtensionTarget,
 >(): Shop {
   return useApi<Target>().shop;
 }
