+++ packages/ui-extensions/src/surfaces/customer-account.ts
@@ -2,7 +2,7 @@ export * from './customer-account/targets';
 export * from './customer-account/extension';
 export * from './customer-account/globals';
 export * from './customer-account/api';
-export * from './customer-account/components';
 
+export * from './checkout/components';
 export * from './checkout/shared';
 export * from './checkout/style';
