+++ packages/ui-extensions/src/surfaces/customer-account/api.ts
@@ -1,16 +1,23 @@
 export type {
+  ApiVersion,
+  BuyerIdentity,
+  Capability,
   CartCost,
   CartLineCost,
   CheckoutSettings,
   Customer,
+  Editor,
+  Extension,
   ExtensionSettings,
   Merchandise,
   ImageDetails,
   Product,
   ProductVariant,
+  PurchasingCompany,
   SelectedOption,
   SessionToken,
   Money,
+  Storage,
   Shop,
   Version,
   Metafield,
@@ -26,9 +33,13 @@ export type {
   CartAutomaticDiscountAllocation,
   CartCustomDiscountAllocation,
   AppliedGiftCard,
+  I18n,
+  I18nTranslate,
   Currency,
   Country,
   Market,
+  Language,
+  Localization,
   DeliveryGroup,
   DeliveryGroupType,
   CartLineReference,
@@ -40,10 +51,6 @@ export type {
   NumberRange,
   Order,
   OrderStatusApi,
-  OrderStatusCompany,
-  OrderStatusLocalization,
-  OrderStatusPurchasingCompany,
-  OrderStatusBuyerIdentity,
 } from './api/order-status/order-status';
 export type {
   Attribute,
@@ -54,30 +61,6 @@ export type {
   SellingPlan,
   StorefrontApiVersion,
   MailingAddress,
-  ApiVersion,
-  Capability,
-  Editor,
-  Extension,
-  I18nTranslate,
-  I18n,
-  Language,
-  Storage,
 } from './api/shared';
 
 export type {CartLineItemApi} from './api/cart-line/cart-line-item';
-
-export type {
-  RendererVersion,
-  StandardApi,
-  FullExtensionNavigation,
-  StandardExtensionNavigation,
-  NavigationCurrentEntryChangeEvent,
-  NavigationHistoryEntry,
-  NavigationOptions,
-  NavigationType,
-  NavigateFunction,
-  BuyerIdentity,
-  PurchasingCompany,
-  Company,
-  Localization,
-} from './api/standard-api/standard-api';
