+++ packages/ui-extensions/src/surfaces/customer-account/api/order-status/order-status.ts
@@ -9,12 +9,130 @@ import type {
   SellingPlan,
   Attribute,
   MailingAddress,
-  I18n,
-  Language,
-  Storage,
 } from '../shared';
 import type {ExtensionTarget} from '../../targets';
-import {Extension} from '../shared';
+import {ApiVersion} from '../../../../shared';
+
+/**
+ * A key-value storage object for extension targets.
+ *
+ * Stored data is only available to this specific app
+ * at this specific extension target.
+ *
+ * The storage backend is implemented with `localStorage` and
+ * should persist across the buyer's checkout session.
+ * However, data persistence isn't guaranteed.
+ */
+export interface Storage {
+  /**
+   * Read and return a stored value by key.
+   *
+   * The stored data is deserialized from JSON and returned as
+   * its original primitive.
+   *
+   * Returns `null` if no stored data exists.
+   */
+  read<T = unknown>(key: string): Promise<T | null>;
+
+  /**
+   * Write stored data for this key.
+   *
+   * The data must be serializable to JSON.
+   */
+  write(key: string, data: any): Promise<void>;
+
+  /**
+   * Delete stored data by key.
+   */
+  delete(key: string): Promise<void>;
+}
+
+/**
+ * The capabilities an extension has access to.
+ *
+ * * [`api_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#api-access): the extension can access the Storefront API.
+ *
+ * * [`network_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#network-access): the extension can make external network calls.
+ *
+ * * [`block_progress`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#block-progress): the extension can block a buyer's progress and the merchant has allowed this blocking behavior.
+ */
+export type Capability = 'api_access' | 'network_access' | 'block_progress';
+
+/**
+ * Meta information about an extension target.
+ */
+export interface Extension<Target extends ExtensionTarget = ExtensionTarget> {
+  /**
+   * The API version that was set in the extension config file.
+   *
+   * @example '2023-04', '2023-07'
+   */
+  apiVersion: ApiVersion;
+
+  /**
+   * The allowed capabilities of the extension, defined
+   * in your [shopify.ui.extension.toml](https://shopify.dev/docs/api/checkout-ui-extensions/configuration) file.
+   *
+   * * [`api_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#api-access): the extension can access the Storefront API.
+   *
+   * * [`network_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#network-access): the extension can make external network calls.
+   *
+   * * [`block_progress`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#block-progress): the extension can block a buyer's progress and the merchant has allowed this blocking behavior.
+   */
+  capabilities: StatefulRemoteSubscribable<Capability[]>;
+
+  /**
+   * Information about the editor where the extension is being rendered.
+   *
+   * The value is undefined if the extension is not rendering in an editor.
+   */
+  editor?: Editor;
+
+  /**
+   * Whether your extension is currently rendered to the screen.
+   *
+   * Shopify might render your extension before it's visible in the UI,
+   * typically to pre-render extensions that will appear on a later step of the
+   * checkout.
+   *
+   * Your extension might also continue to run after the buyer has navigated away
+   * from where it was rendered. The extension continues running so that
+   * your extension is immediately available to render if the buyer navigates back.
+   */
+  rendered: StatefulRemoteSubscribable<boolean>;
+
+  /**
+   * The URL to the script that started the extension target.
+   */
+  scriptUrl: string;
+
+  /**
+   * The identifier that specifies where in Shopify’s UI your code is being
+   * injected. This will be one of the targets you have included in your
+   * extension’s configuration file.
+   *
+   * @example 'customer-account.order-status.block.render'
+   * @see https://shopify.dev/docs/api/checkout-ui-extensions/unstable/extension-targets-overview
+   * @see https://shopify.dev/docs/apps/app-extensions/configuration#targets
+   */
+  target: Target;
+
+  /**
+   * The published version of the running extension target.
+   *
+   * For unpublished extensions, the value is `undefined`.
+   *
+   * @example 3.0.10
+   */
+  version?: string;
+}
+
+export interface Editor {
+  /**
+   * Indicates whether the extension is rendering in the checkout editor.
+   */
+  type: 'checkout';
+}
 
 /**
  * Metadata associated with the checkout.
@@ -116,9 +234,88 @@ export interface AppMetafieldEntry {
   metafield: AppMetafield;
 }
 
-export type ApiVersion = '2023-04' | '2023-07' | 'unstable';
+export type {ApiVersion} from '../../../../shared';
 
 export type Version = string;
+
+/**
+ * This returns a translated string matching a key in a locale file.
+ *
+ * @example translate("banner.title")
+ */
+export type I18nTranslate<ReplacementType = string> = (
+  key: string,
+  options?: {[placeholderKey: string]: ReplacementType | string | number},
+) => ReplacementType extends string | number
+  ? string
+  : (string | ReplacementType)[];
+
+export interface I18n {
+  /**
+   * Returns a localized number.
+   *
+   * This function behaves like the standard `Intl.NumberFormat()`
+   * with a style of `decimal` applied. It uses the buyer's locale by default.
+   *
+   * @param options.inExtensionLocale - if true, use the extension's locale
+   */
+  formatNumber: (
+    number: number | bigint,
+    options?: {inExtensionLocale?: boolean} & Intl.NumberFormatOptions,
+  ) => string;
+
+  /**
+   * Returns a localized currency value.
+   *
+   * This function behaves like the standard `Intl.NumberFormat()`
+   * with a style of `currency` applied. It uses the buyer's locale by default.
+   *
+   * @param options.inExtensionLocale - if true, use the extension's locale
+   */
+  formatCurrency: (
+    number: number | bigint,
+    options?: {inExtensionLocale?: boolean} & Intl.NumberFormatOptions,
+  ) => string;
+
+  /**
+   * Returns a localized date value.
+   *
+   * This function behaves like the standard `Intl.DateTimeFormatOptions()` and uses
+   * the buyer's locale by default. Formatting options can be passed in as
+   * options.
+   *
+   * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat0
+   * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat#using_options
+   *
+   * @param options.inExtensionLocale - if true, use the extension's locale
+   */
+  formatDate: (
+    date: Date,
+    options?: {inExtensionLocale?: boolean} & Intl.DateTimeFormatOptions,
+  ) => string;
+
+  /**
+   * Returns translated content in the buyer's locale,
+   * as supported by the extension.
+   *
+   * - `options.count` is a special numeric value used in pluralization.
+   * - The other option keys and values are treated as replacements for interpolation.
+   * - If the replacements are all primitives, then `translate()` returns a single string.
+   * - If replacements contain UI components, then `translate()` returns an array of elements.
+   */
+  translate: I18nTranslate;
+}
+
+export interface Language {
+  /**
+   * The BCP-47 language tag. It may contain a dash followed by an ISO 3166-1 alpha-2 region code.
+   *
+   * @example 'en' for English, or 'en-US' for English local to United States.
+   * @see https://en.wikipedia.org/wiki/IETF_language_tag
+   * @see https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2
+   */
+  isoCode: string;
+}
 
 export interface Currency {
   /**
@@ -148,7 +345,7 @@ export interface Market {
   handle: string;
 }
 
-export interface OrderStatusLocalization {
+export interface Localization {
   /**
    * The currency that the buyer sees for money amounts in the checkout.
    */
@@ -226,7 +423,7 @@ export interface OrderStatusApi<Target extends ExtensionTarget> {
    *
    * {% include /apps/checkout/privacy-icon.md %} Requires access to [protected customer data](/docs/apps/store/data-protection/protected-customer-data).
    */
-  buyerIdentity?: OrderStatusBuyerIdentity;
+  buyerIdentity?: BuyerIdentity;
 
   /**
    * Settings applied to the buyer's checkout.
@@ -287,7 +484,7 @@ export interface OrderStatusApi<Target extends ExtensionTarget> {
    * [`i18n`](https://shopify.dev/docs/api/checkout-ui-extensions/apis/standardapi#properties-propertydetail-i18n)
    * object instead.
    */
-  localization: OrderStatusLocalization;
+  localization: Localization;
 
   /**
    * The metafields that apply to the current order. The actual resource
@@ -394,7 +591,7 @@ export interface SessionToken {
   get(): Promise<string>;
 }
 
-export interface OrderStatusBuyerIdentity {
+export interface BuyerIdentity {
   /**
    * The buyer's customer account. The value is undefined if the buyer isn’t a
    * known customer for this shop or if they haven't logged in yet.
@@ -426,30 +623,28 @@ export interface OrderStatusBuyerIdentity {
    *
    * {% include /apps/checkout/privacy-icon.md %} Requires access to [protected customer data](/docs/apps/store/data-protection/protected-customer-data).
    */
-  purchasingCompany: StatefulRemoteSubscribable<
-    OrderStatusPurchasingCompany | undefined
-  >;
+  purchasingCompany: StatefulRemoteSubscribable<PurchasingCompany | undefined>;
 }
 
 /**
  * Information about a company that the business customer is purchasing on behalf of.
  */
-export interface OrderStatusPurchasingCompany {
+export interface PurchasingCompany {
   /**
    * Includes information of the company that the business customer is purchasing on behalf of.
    *
    * {% include /apps/checkout/privacy-icon.md %} Requires access to [protected customer data](/docs/apps/store/data-protection/protected-customer-data).
    */
-  company: OrderStatusCompany;
+  company: Company;
   /**
    * Includes information of the company location that the business customer is purchasing on behalf of.
    *
    * {% include /apps/checkout/privacy-icon.md %} Requires access to [protected customer data](/docs/apps/store/data-protection/protected-customer-data).
    */
-  location: OrderStatusCompanyLocation;
+  location: CompanyLocation;
 }
 
-export interface OrderStatusCompany {
+export interface Company {
   /**
    * The company ID.
    *
@@ -470,7 +665,7 @@ export interface OrderStatusCompany {
   externalId?: string;
 }
 
-export interface OrderStatusCompanyLocation {
+export interface CompanyLocation {
   /**
    * The company location ID.
    *
@@ -529,26 +724,6 @@ export interface Shop {
 }
 
 export interface CartCost {
-  /**
-   * A `Money` value representing the subtotal value of the items in the cart at the current
-   * step of checkout.
-   */
-  subtotalAmount: StatefulRemoteSubscribable<Money>;
-
-  /**
-   * A `Money` value representing the total shipping a buyer can expect to pay at the current
-   * step of checkout. This value includes shipping discounts. Returns undefined if shipping
-   * has not been negotiated yet, such as on the information step.
-   */
-  totalShippingAmount: StatefulRemoteSubscribable<Money | undefined>;
-
-  /**
-   * A `Money` value representing the total tax a buyer can expect to pay at the current
-   * step of checkout or the total tax included in product and shipping prices. Returns
-   * undefined if taxes are unavailable.
-   */
-  totalTaxAmount: StatefulRemoteSubscribable<Money | undefined>;
-
   /**
    * A `Money` value representing the minimum a buyer can expect to pay at the current
    * step of checkout. This value excludes amounts yet to be negotiated. For example,
