+++ packages/ui-extensions/src/surfaces/customer-account/api/shared.ts
@@ -1,214 +1,3 @@
-import type {ExtensionTarget} from '../targets';
-import type {StatefulRemoteSubscribable} from '@remote-ui/async-subscription';
-import {ApiVersion} from '../../../shared';
-
-/**
- * A key-value storage object for extension targets.
- *
- * Stored data is only available to this specific app
- * but can be shared across multiple extension targets.
- *
- * The storage backend is implemented with `localStorage` and
- * should persist for ... days
- * However, data persistence isn't guaranteed.
- */
-export interface Storage {
-  /**
-   * Read and return a stored value by key.
-   *
-   * The stored data is deserialized from JSON and returned as
-   * its original primitive.
-   *
-   * Returns `null` if no stored data exists.
-   */
-  read<T = unknown>(key: string): Promise<T | null>;
-
-  /**
-   * Write stored data for this key.
-   *
-   * The data must be serializable to JSON.
-   */
-  write(key: string, data: any): Promise<void>;
-
-  /**
-   * Delete stored data by key.
-   */
-  delete(key: string): Promise<void>;
-}
-
-export interface Language {
-  /**
-   * The BCP-47 language tag. It may contain a dash followed by an ISO 3166-1 alpha-2 region code.
-   *
-   * @example 'en' for English, or 'en-US' for English local to United States.
-   * @see https://en.wikipedia.org/wiki/IETF_language_tag
-   * @see https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2
-   */
-  isoCode: string;
-}
-
-/**
- * This defines the i18n.translate() signature.
- */
-export interface I18nTranslate {
-  /**
-   * This returns a translated string matching a key in a locale file.
-   *
-   * @example translate("banner.title")
-   */
-  <ReplacementType = string>(
-    key: string,
-    options?: {[placeholderKey: string]: ReplacementType | string | number},
-  ): ReplacementType extends string | number
-    ? string
-    : (string | ReplacementType)[];
-}
-
-export interface I18n {
-  /**
-   * Returns a localized number.
-   *
-   * This function behaves like the standard `Intl.NumberFormat()`
-   * with a style of `decimal` applied. It uses the buyer's locale by default.
-   *
-   * @param options.inExtensionLocale - if true, use the extension's locale
-   */
-  formatNumber: (
-    number: number | bigint,
-    options?: {inExtensionLocale?: boolean} & Intl.NumberFormatOptions,
-  ) => string;
-
-  /**
-   * Returns a localized currency value.
-   *
-   * This function behaves like the standard `Intl.NumberFormat()`
-   * with a style of `currency` applied. It uses the buyer's locale by default.
-   *
-   * @param options.inExtensionLocale - if true, use the extension's locale
-   */
-  formatCurrency: (
-    number: number | bigint,
-    options?: {inExtensionLocale?: boolean} & Intl.NumberFormatOptions,
-  ) => string;
-
-  /**
-   * Returns a localized date value.
-   *
-   * This function behaves like the standard `Intl.DateTimeFormatOptions()` and uses
-   * the buyer's locale by default. Formatting options can be passed in as
-   * options.
-   *
-   * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat0
-   * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat#using_options
-   *
-   * @param options.inExtensionLocale - if true, use the extension's locale
-   */
-  formatDate: (
-    date: Date,
-    options?: {inExtensionLocale?: boolean} & Intl.DateTimeFormatOptions,
-  ) => string;
-
-  /**
-   * Returns translated content in the buyer's locale,
-   * as supported by the extension.
-   *
-   * - `options.count` is a special numeric value used in pluralization.
-   * - The other option keys and values are treated as replacements for interpolation.
-   * - If the replacements are all primitives, then `translate()` returns a single string.
-   * - If replacements contain UI components, then `translate()` returns an array of elements.
-   */
-  translate: I18nTranslate;
-}
-
-/**
- * The capabilities an extension has access to.
- *
- * * [`api_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#api-access): the extension can access the Storefront API.
- *
- * * [`network_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#network-access): the extension can make external network calls.
- *
- * * [`block_progress`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#block-progress): the extension can block a buyer's progress and the merchant has allowed this blocking behavior.
- */
-export type Capability = 'api_access' | 'network_access' | 'block_progress';
-
-/**
- * Meta information about an extension target.
- */
-export interface Extension<Target extends ExtensionTarget = ExtensionTarget> {
-  /**
-   * The API version that was set in the extension config file.
-   *
-   * @example '2023-04', '2023-07'
-   */
-  apiVersion: ApiVersion;
-
-  /**
-   * The allowed capabilities of the extension, defined
-   * in your [shopify.ui.extension.toml](https://shopify.dev/docs/api/checkout-ui-extensions/configuration) file.
-   *
-   * * [`api_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#api-access): the extension can access the Storefront API.
-   *
-   * * [`network_access`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#network-access): the extension can make external network calls.
-   *
-   * * [`block_progress`](https://shopify.dev/docs/api/checkout-ui-extensions/configuration#block-progress): the extension can block a buyer's progress and the merchant has allowed this blocking behavior.
-   */
-  capabilities: StatefulRemoteSubscribable<Capability[]>;
-
-  /**
-   * Information about the editor where the extension is being rendered.
-   *
-   * The value is undefined if the extension is not rendering in an editor.
-   */
-  editor?: Editor;
-
-  /**
-   * Whether your extension is currently rendered to the screen.
-   *
-   * Shopify might render your extension before it's visible in the UI,
-   * typically to pre-render extensions that will appear on a later step of the
-   * checkout.
-   *
-   * Your extension might also continue to run after the buyer has navigated away
-   * from where it was rendered. The extension continues running so that
-   * your extension is immediately available to render if the buyer navigates back.
-   */
-  rendered: StatefulRemoteSubscribable<boolean>;
-
-  /**
-   * The URL to the script that started the extension target.
-   */
-  scriptUrl: string;
-
-  /**
-   * The identifier that specifies where in Shopify’s UI your code is being
-   * injected. This will be one of the targets you have included in your
-   * extension’s configuration file.
-   *
-   * @example 'customer-account.order-status.block.render'
-   * @see https://shopify.dev/docs/api/checkout-ui-extensions/unstable/extension-targets-overview
-   * @see https://shopify.dev/docs/apps/app-extensions/configuration#targets
-   */
-  target: Target;
-
-  /**
-   * The published version of the running extension target.
-   *
-   * For unpublished extensions, the value is `undefined`.
-   *
-   * @example 3.0.10
-   */
-  version?: string;
-}
-
-export interface Editor {
-  /**
-   * Indicates whether the extension is rendering in the checkout editor.
-   */
-  type: 'checkout';
-}
-
-export type {ApiVersion} from '../../../shared';
-
 export type ValueOrPromise<T> = T extends PromiseLike<any> ? T : T | Promise<T>;
 
 export type CurrencyCode =
