--- packages/ui-extensions/src/surfaces/customer-account/components/CustomerAccountAction/CustomerAccountAction.ts
@@ -1,22 +1,0 @@
-import {createRemoteComponent} from '@remote-ui/core';
-import type {RemoteFragment} from '@remote-ui/core';
-
-export interface CustomerAccountActionProps {
-  /**
-   * Sets the title of the Action container.
-   */
-  title: string;
-  /**
-   * Sets the Primary action button of the container. This component must a button component.
-   */
-  primaryAction?: RemoteFragment;
-  /**
-   * Sets the Secondary action button of the container. This component must a button component.
-   */
-  secondaryAction?: RemoteFragment;
-}
-
-export const CustomerAccountAction = createRemoteComponent<
-  'CustomerAccountAction',
-  CustomerAccountActionProps
->('CustomerAccountAction');
