--- packages/ui-extensions/src/surfaces/customer-account/components/PaymentIcon/index.ts
@@ -1,2 +1,0 @@
-export {PaymentIcon} from './PaymentIcon';
-export type {PaymentIconProps, PaymentIconSource} from './PaymentIcon';
