--- packages/ui-extensions/src/surfaces/customer-account/components/PolicyModal/index.ts
@@ -1,2 +1,0 @@
-export {PolicyModal} from './PolicyModal';
-export type {PolicyModalProps} from './PolicyModal';
