--- packages/ui-extensions/src/surfaces/customer-account/components/index.ts
@@ -1,6 +1,0 @@
-export * from './Card';
-export * from './CustomerAccountAction';
-export * from './PaymentIcon';
-export * from './PolicyModal';
-export * from './Thumbnail';
-export * from './shared-checkout-components';
