--- packages/ui-extensions/src/surfaces/customer-account/components/shared-checkout-components.ts
@@ -1,105 +1,0 @@
-export {
-  Banner,
-  BlockLayout,
-  BlockSpacer,
-  BlockStack,
-  Button,
-  Checkbox,
-  Choice,
-  ChoiceList,
-  DateField,
-  DatePicker,
-  Disclosure,
-  Divider,
-  Form,
-  Grid,
-  GridItem,
-  Heading,
-  HeadingGroup,
-  Icon,
-  Image,
-  InlineLayout,
-  InlineStack,
-  InlineSpacer,
-  Link,
-  List,
-  ListItem,
-  Pressable,
-  Select,
-  SkeletonImage,
-  SkeletonText,
-  SkeletonTextBlock,
-  Spinner,
-  Stepper,
-  Tag,
-  Text,
-  TextBlock,
-  TextField,
-  View,
-  Modal,
-  Popover,
-  Tooltip,
-  ScrollView,
-} from '../../checkout/components';
-
-export type {
-  BannerProps,
-  BlockLayoutProps,
-  BlockSpacerProps,
-  BlockStackProps,
-  ButtonProps,
-  CheckboxProps,
-  ChoiceProps,
-  ChoiceListProps,
-  DatePickerProps,
-  DateFieldProps,
-  DisclosureProps,
-  DividerProps,
-  FormProps,
-  GridProps,
-  GridItemSize,
-  GridItemProps,
-  HeadingProps,
-  HeadingGroupProps,
-  IconProps,
-  ImageProps,
-  InlineLayoutProps,
-  InlineStackProps,
-  InlineSpacerProps,
-  LinkProps,
-  ListProps,
-  ListItemProps,
-  MaybeConditionalStyle,
-  MaybeShorthandProperty,
-  ModalProps,
-  PopoverProps,
-  PressableProps,
-  SelectProps,
-  SkeletonImageProps,
-  SkeletonTextProps,
-  SkeletonTextBlockProps,
-  SpinnerProps,
-  StepperProps,
-  TagProps,
-  TextProps,
-  TextBlockProps,
-  TextFieldProps,
-  ViewProps,
-  ViewPosition,
-  ViewPositionType,
-  ViewTranslate,
-  ViewCoordinate,
-  AccessibilityRole,
-  ViewLikeAccessibilityRole,
-  NonPresentationalAccessibilityRole,
-  Autocomplete,
-  AutocompleteField,
-  AutocompleteGroup,
-  Status,
-  Size,
-  Spacing,
-  Alignment,
-  Appearance,
-  ScrollViewProps,
-  TooltipProps,
-} from '../../checkout';
