+++ packages/ui-extensions/src/surfaces/customer-account/extension.ts
@@ -19,7 +19,3 @@ export const extension =
  * @deprecated This is deprecated, use `extension` instead.
  */
 export const extend = extension;
-
-export interface RunExtension<Api, Return> {
-  (api: Api): Promise<Return>;
-}
